Fix detection of invalid packages
During the merging phase of a package, emerge removes the corresponding
directory in /var/db/pkg/, creates a new directory prefixed with
"-MERGING-", adds the required files to it and moves it back into place.

If haskell-updater is run during a emerge merge, it wrongly assumes
there's a "<cat>/-MERGING-<package>" package due to this behaviour.

Therefore, let haskell-updater ignore such directories.

Signed-off-by: Luis Ressel <aranea@aixah.de>