haskell-updater: rerun upgrader passes as long as there is progress
The primary motivation here is to handle situation when
haskell-updater run causes more package breakage not visible
before haskell-updater run.

The hypothetical example is the following setup:
- dev-haskell/text-1 installed
- dev-haskell/libfoo-0 installed (depends on dev-haskell/text-1)
- dev-haskell/libbar-0 installed (depends on dev-haskell/libfoo-0)
- dev-haskell/libbroken-0 installed and broken (depends on dev-haskell/text-1)
- dev-haskell/text-2 available but not yet installed.

In this case haskell-updater would catch only 'dev-haskell/libbroken'
as broken and would reschedule it for rebuild. As haskell-updater
uses --deep option that causes the following rebuilds:
- dev-haskell/libbroken-0 (as haskell-updater detects it as broken)
- dev-haskell/text-2 (as it's an upgrade dev-haskell/text-1,
                      dependency of 'libbroken')
  - dev-haskell/libfoo-0 (as it's caused by 'text' subslot change)

If the above steps succeed user is left with one package
freshly broken by upgrader run: 'dev-haskell/libbar-0'

Rerun of haskell-updater would fix it. This change does precisely that.

After this change haskell-updater runs as long as there is packages to rebuild.
To avoid rebuild loops which can be caused broken dependencies of failing
to build packages haskell-updater makes sure each run is unique.

Closes: https://github.com/gentoo-haskell/haskell-updater/issues/7
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>