Use --usepkg-exclude instead of --usepkg=n
For Portage only. When building the emerge command, add relevant
packages using --usepkg-exclude to make sure needed packages are
rebuilt. For normal mode, this is the list of target packages. For
reinstall-atoms mode, this is the list of all currently installed
packages.

Add AllPkgs as a parameter to RAModeState to signify that the set of all
packages is always needed, as this is what is used to build the
--usepkg-exclude list.

Signed-off-by: hololeap <hololeap@protonmail.com>