Adjust loop logic for reinstall-atoms mode
When running in reinstall-atoms mode, use adjusted logic to control the
looping mechanism: If emerge runs and no changes are made to the
broken target list, check the previous run's exit code. If it was a
success, exit successfully. If it was a failure, exit with a failure
message. Otherwise, continue running emerge. This logic does not apply
to --target=all as this mode does not have a way to track progress
between emerge runs.

Remove the Maybe from RAModeState as an empty package list (signified by
Nothing) does not necessarily mean a success.

Use a Sequence to track history as this is more natural than the
`Map (Set Package) Int` that it is replacing. Store the exit code along
with the target package list in the history.

Signed-off-by: hololeap <hololeap@protonmail.com>