Main: Add note about dep resolution failure
Add a note to the command line output for the user to check for
dependency resolution failure in the package manager output. This
usually comes up with the UntilNoChange loop type (e.g. --world mode)
and can be guessed by haskell-updater if it failed on its first attempt.

Dependency resolution failure falls outside the scope of
haskell-updater and it is generally the responsibility of the user to
fix, assuming there is not an error in the repository.

Signed-off-by: hololeap <hololeap@protonmail.com>